Changes, Additions
Changed where the module load function is located for extended use of
vars array.

Changed the way modules will be installed. Once a module is uploaded and
activated, you must visit the modules page to have it installed. Once
this is complete you can then use the module. This also adds a variable
to the generation modifiers that indicates that a module is installed.